Retarget OQS-OpenSSL repo in build.py to fix #7
At some point, OQS deleted the branch we were using and rebuilt the 1.0.2 fork branch. This means the commit we were using as our snapshot no longer exists and we can't check it out specifically.

Use kevinmkane's fork on GitHub, which had a copy of this repo at this time, so we can stay on that same snapshot for now. We'll return to the official repo when we do our update to use the 1.1 branch.

As an aside, change run_command in build.py to raise an exception on a nonzero return value from the command for better catching of failures like this one. (The checkout command had been failing but this was ignored.)