Don't use submodules for the copies of OQS-OpenSSL we build on Windows
All three copies of OQS-OpenSSL were submodules: the one subtree we use for Linux,
and the two subtrees on Windows. The latter two used the former as their origin, which
is proving problematic.

To simplify things, only have one submodule in openvpn\build\repos\openssl-oqs, and have
all development happen in there. In the Windows build we'll make copies of this tree for
the builds.