Remove building on Windows from build.py
liboqs and OQS-OpenSSL can now be cross-compiled on Linux for Windows. Remove all support for building them on Windows, since this saves us the awkward step of building it on Windows, copying the DLLs over to a Linux host, and continuing the build there. Now everything is built on Linux, including cross-compiled OpenSSL and OpenVPN binaries for Windows.