Conform to OpenVPN behavior for letting OpenSSL choose suitable defaults
A previous change in the OpenVPN submodule tried to choose p256-sikep434 as a default if no ecdh-curve directive was present, and this worked because of a bug in OpenVPN that was fixed in v2.4.9. Instead of working around this, we return to OpenVPN's intended behavior, since the OQS fork of OpenSSL should eventually take charge of picking suitable defaults to only be post-quantum. Expand the comments in the sample configuration files to make it clear the ecdh-curve directive must be specified at least on the server to guarantee a post-=quantum key exchange.